+++ ttsrv/listener.go
@@ -16,7 +16,7 @@ func NewListener() *Listener {
 		Up:            make(chan struct{}, 0),
 		AcceptTimeout: 100 * time.Millisecond,
 		Logger:        log.New(os.Stdout, "tcp ", log.Flags()),
-		AddConnection: func(context.Context, Remote) { /*noop*/ },
+		AddConnection: func(context.Context, Connection) { /*noop*/ },
 	}
 }
 
@@ -35,7 +35,7 @@ type Listener struct {
 	AcceptTimeout time.Duration
 
 	// AddConnection handles new remote connections
-	AddConnection func(context.Context, Remote)
+	AddConnection func(context.Context, Connection)
 	*log.Logger
 }